Moved KafkaPluginNonBlocking to separate module.
